requirements.txt
tornado
python-escpos
